Refactored the OpenTelemetry tools
