+++ CloudDebugger/Features/OpenTelemetry/Index.cshtml
@@ -17,12 +17,12 @@ Tools related to OpenTelemetry.
 <br />
 
 
-<h4>OpenTelemetry Logs</h4>
-<a href="/OpenTelemetryLog/ViewTraces">View trace data</a>
+<h4>OpenTelemetry Data</h4>
+<a href="/OpenTelemetryTracesViewer/ViewTraces">View trace data</a>
 <br />
-<a href="/OpenTelemetryLog/ViewMetrics">View metrics data</a>
+<a href="/OpenTelemetryMetricsViewer/ViewMetrics">View metrics data</a>
 <br />
-<a href="/OpenTelemetryLog/ViewLogs">View Log data</a>
+<a href="/OpenTelemetryLogViewer/ViewLogs">View Log data</a>
 <br />
 <br />
 <br />
