+++ CloudDebugger/Features/OpenTelemetryLogViewer/ViewLogs.cshtml
@@ -1,4 +1,4 @@
-﻿@model CloudDebugger.Features.OpenTelemetryLog.ViewLogsModel
+﻿@model CloudDebugger.Features.OpenTelemetryLogViewer.ViewLogsModel
 @{
     ViewData["Title"] = "Cloud Debugger - View OpenTelemetry Logs";
 }
