+++ CloudDebugger/Features/OpenTelemetryMetricsViewer/ViewMetrics.cshtml
@@ -1,4 +1,4 @@
-﻿@model CloudDebugger.Features.OpenTelemetryLog.ViewMetricsModel
+﻿@model CloudDebugger.Features.OpenTelemetryMetricsViewer.ViewMetricsModel
 @{
     ViewData["Title"] = "Cloud Debugger - View OpenTelemetry Metrics";
 }
