+++ CloudDebugger/Features/OpenTelemetryTracesViewer/ViewTraces.cshtml
@@ -1,11 +1,11 @@
-﻿@model CloudDebugger.Features.OpenTelemetryLog.ViewTracesModel
+﻿@model CloudDebugger.Features.OpenTelemetryTracesViewer.ViewTracesModel
 @{
     ViewData["Title"] = "Cloud Debugger - View OpenTelemetry Traces";
 }
 
 <h5>View OpenTelemetry Traces</h5>
 This page will display the latest OpenTelemetry traces.<br />
-<a href="/OpenTelemetry/index">Back</a>&nbsp;&nbsp;<a href="/OpenTelemetryLog/ClearTraceLog">Clear the log</a>
+<a href="/OpenTelemetry/index">Back</a>&nbsp;&nbsp;<a href="/OpenTelemetryTracesViewer/ClearTraceLog">Clear the log</a>
 <br />
 <br />
 It will ignore the following traces:
