+++ CloudDebugger/Features/OpenTelemetryTracesViewer/ViewTracesModel.cs
@@ -1,4 +1,4 @@
-﻿namespace CloudDebugger.Features.OpenTelemetryLog;
+﻿namespace CloudDebugger.Features.OpenTelemetryTracesViewer;
 
 public class ViewTracesModel
 {
